requirements.txt
fastapi
uvicorn
nest-asyncio
pyngrok
paddleocr
opencv-python-headless
transformers
torch
sentencepiece
python-multipart
pydantic
PyPDF2
paddlepaddle
pillow